Add /drive/files/attached-notes
I decided to split up the drive group and create the first
"sub-service". It works the same way the normal services, they could
call any endpoints, no restrictions. Maybe later all request will be
just processed their parents. It shouldn't be a big problem, just add a
new hander that prepend the root endpoint on the called endpoint path.

But for now, it's not there, later it can be changed under the hood, it
does not change the interface from an external called.