Implement Drive (#29)
Implement Drive

I decided to split up the drive group and create the first
"sub-service". It works the same way the normal services, they could
call any endpoints, no restrictions. Maybe later all request will be
just processed their parents. It shouldn't be a big problem, just add a
new handler that prepend the root endpoint on the called endpoint path.

But for now, it's not there, later it can be changed under the hood, it
does not change the interface for an external caller.

= New request type for multipart requests.

BaseRequest is kind of lame now, as it's basically a JSONRequest...
But I don't want to change that everywhere now. External callers
will don't see that so we can change it later when everything is in
place and we know what works and what not.

UploadFromURL can be called, but I don't recommend it.
That's why I added a CreateFromURL function too, it
downloads the file and use the normal create endpoint.

= First try on tests.

I just wanted to add some guidance how can we test our stuff.
I don't think we have anything to test really as we only Marshal
and Unmarshal []byte{} content.