Admin/Queue endpoints
All Admin/Qeueue endpoints

Ability to create dynamic mock client with MakeMockClient. The
mock client checks if function is there or not. If it's there
it calls the function and uses the returned string as filename.
If the returned string is empty, it falls back to the ResponseFile
field.

Handle 404 as an error. If an endpoint call ends up with a 404 response,
Client will return with an EndpointNotFound error.