Fix formRequest and update linter
The original PR from @itsalltoast would fix on service layer, but in the
future we can face the same issue when we add new endpoints, so decided
to fix it in the "core" with checking if it's a pointer or not.

Now we have tests to cover formRequest endpoints. Both create file and
create file from URL have test.

As an addition, linters are up to date and all offenses are fixed.
original PR: https://github.com/yitsushi/go-misskey/pull/82/files

resolves #81