chore(deps): Update Go version to 1.18 (#86)
Go 1.18 is out. I know it's not well spread yet, but in this form,
everything works as before and I have no short term plans to implement
things that requires 1.18 features yet (like generics).

I was shocked when I saw the version was still Go 1.15.

If you have any issues with this version bump, please file an issue and
we can discuss about it.